Разбил код на модули и настроил eslint
