+++ .gitignore
@@ -13,3 +13,7 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+
+# Собранный JS-код проекта
+bundle.js
+bundle.js.map
