+++ public/index.html
@@ -144,6 +144,6 @@
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
